refactor: Extract hook
