+++ apps/web/src/studio/hooks/useBridgeUrl.ts
@@ -2,7 +2,7 @@ import { useLocalStorage } from '@mantine/hooks';
 import { useLocation } from 'react-router-dom';
 import { TUNNEL_URL_LOCAL_STORAGE_KEY } from '../../api/bridge/utils';
 import { useEnvironment } from '../../hooks/useEnvironment';
-import { isStudioRoute } from './isStudioRoute';
+import { isStudioRoute } from '../utils/isStudioRoute';
 
 export const useBridgeUrl = () => {
   const { pathname } = useLocation();
@@ -11,7 +11,7 @@ export const useBridgeUrl = () => {
   const [bridgeUrl, setBridgeUrl] = useLocalStorage({ key: TUNNEL_URL_LOCAL_STORAGE_KEY });
 
   return {
-    bridgeUrl: isStudioRoute(pathname) ? bridgeUrl : environment?.echo?.url,
+    bridgeUrl: isStudioRoute(pathname) ? bridgeUrl : environment?.echo?.url ?? '',
     setBridgeUrl,
     isLoading,
   };
