+++ apps/web/src/studio/utils/index.ts
@@ -1,2 +1,1 @@
 export * from './isStudioRoute';
-export * from './useBridgeUrl';
